+++ tic-tac-toe-app/package.json
@@ -6,7 +6,6 @@
     "@testing-library/jest-dom": "^5.15.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^26.0.24",
     "@types/node": "^12.20.37",
     "@types/react": "^17.0.37",
     "@types/react-dom": "^17.0.11",
@@ -43,6 +42,7 @@
     ]
   },
   "devDependencies": {
-    "@playwright/test": "^1.17.0"
+    "@playwright/test": "^1.17.0",
+    "@types/jest": "^27.4.0"
   }
 }
