+++ tic-tac-toe-app/src/components/Game.tsx
@@ -1,6 +1,7 @@
 import { FC } from 'react';
 import { useState } from 'react';
 import Board from './Board'
+import { calculateWinner } from './logic'
 
 const Game: FC = () => {
   const [history, setHistory] = useState<{squares: Array<string | null>}[]>(
@@ -25,10 +26,10 @@ const Game: FC = () => {
     setStepNumber(step)
     setXIsNext((step % 2) === 0,)
   }
-  const History = history;
-  const current = History[stepNumber];
+  // const History = history;
+  const current = history[stepNumber];
   const winner = calculateWinner(current.squares);
-  const moves = History.map((step, move) => {
+  const moves = history.map((step, move) => {
     const desc = move ?
       'Go to move #' + move :
       'Go to game start';
@@ -60,24 +61,5 @@ const Game: FC = () => {
   );
 }
 
-const calculateWinner = (squares: Array<string | null>) => {
-  const lines = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6],
-  ];
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return squares[a];
-    }
-  }
-  return null;
-}
 
 export default Game
