+++ tic-tac-toe-app/src/components/__test__/Game.test.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+
+import Game from '../Game';
+
+describe('Game', () => {
+  test('renders App component', () => {
+    render(
+      <Game />
+      );
+  });
+});
